Enable Solhint (#10869)
* enable solhint for /shared/ folder

add others to ignore file
add solhint to CICD

* add chainlink-solidity ruleset

* reduce ignore scope & fix interface folder issues

* rm unused imports

* clean /libraries

* set no-unused-import to error

* allow 968 warnings, the number we currently have

* fix second layer imports

* contracts/v0.8: fix solhint warnings for vrf contracts (#10875)

* rm no-inline-assembly rule

---------

+++ contracts/src/v0.8/dev/vrf/testhelpers/VRFCoordinatorV2PlusUpgradedVersion.sol
@@ -1,17 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.4;
 
-import "../../../shared/interfaces/LinkTokenInterface.sol";
-import "../../../interfaces/BlockhashStoreInterface.sol";
-import "../../../interfaces/TypeAndVersionInterface.sol";
-import "../../interfaces/IVRFCoordinatorV2Plus.sol";
-import "../../../vrf/VRF.sol";
-import "../VRFConsumerBaseV2Plus.sol";
-import "../../../ChainSpecificUtil.sol";
-import "../SubscriptionAPI.sol";
-import "../libraries/VRFV2PlusClient.sol";
-import "../../interfaces/IVRFCoordinatorV2PlusMigration.sol";
-import "../../../vendor/openzeppelin-solidity/v4.7.3/contracts/utils/structs/EnumerableSet.sol";
+import {BlockhashStoreInterface} from "../../../interfaces/BlockhashStoreInterface.sol";
+// solhint-disable-next-line no-unused-import
+import {IVRFCoordinatorV2Plus, IVRFSubscriptionV2Plus} from "../../interfaces/IVRFCoordinatorV2Plus.sol";
+import {VRF} from "../../../vrf/VRF.sol";
+import {VRFConsumerBaseV2Plus, IVRFMigratableConsumerV2Plus} from "../VRFConsumerBaseV2Plus.sol";
+import {ChainSpecificUtil} from "../../../ChainSpecificUtil.sol";
+import {SubscriptionAPI} from "../SubscriptionAPI.sol";
+import {VRFV2PlusClient} from "../libraries/VRFV2PlusClient.sol";
+import {IVRFCoordinatorV2PlusMigration} from "../../interfaces/IVRFCoordinatorV2PlusMigration.sol";
+import {EnumerableSet} from "../../../vendor/openzeppelin-solidity/v4.7.3/contracts/utils/structs/EnumerableSet.sol";
 
 contract VRFCoordinatorV2PlusUpgradedVersion is
   VRF,
@@ -21,6 +20,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
 {
   using EnumerableSet for EnumerableSet.UintSet;
   /// @dev should always be available
+  // solhint-disable-next-line chainlink-solidity/prefix-immutable-variables-with-i
   BlockhashStoreInterface public immutable BLOCKHASH_STORE;
 
   // Set this maximum to 200 to give us a 56 block window to fulfill
@@ -291,6 +291,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
     return requestId;
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function computeRequestId(
     bytes32 keyHash,
     address sender,
@@ -305,8 +306,8 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
    * @dev calls target address with exactly gasAmount gas and data as calldata
    * or reverts if at least gasAmount gas is not available.
    */
+  // solhint-disable-next-line chainlink-solidity/prefix-private-functions-with-underscore
   function callWithExactGas(uint256 gasAmount, address target, bytes memory data) private returns (bool success) {
-    // solhint-disable-next-line no-inline-assembly
     assembly {
       let g := gas()
       // Compute g -= GAS_FOR_CALL_EXACT_CHECK and check for underflow
@@ -341,6 +342,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
     uint256 randomness;
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function getRandomnessFromProof(
     Proof memory proof,
     RequestCommitment memory rc
@@ -444,6 +446,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
     }
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function calculatePaymentAmount(
     uint256 startGas,
     uint256 gasAfterPaymentCalculation,
@@ -468,6 +471,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
       );
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function calculatePaymentAmountNative(
     uint256 startGas,
     uint256 gasAfterPaymentCalculation,
@@ -485,6 +489,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
   }
 
   // Get the amount of gas used for fulfillment
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function calculatePaymentAmountLink(
     uint256 startGas,
     uint256 gasAfterPaymentCalculation,
@@ -508,6 +513,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
     return uint96(paymentNoFee + fee);
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-private-functions-with-underscore
   function getFeedData() private view returns (int256) {
     uint32 stalenessSeconds = s_config.stalenessSeconds;
     bool staleFallback = stalenessSeconds > 0;
@@ -615,6 +621,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
     uint96 nativeBalance;
   }
 
+  // solhint-disable-next-line chainlink-solidity/prefix-internal-functions-with-underscore
   function isTargetRegistered(address target) internal view returns (bool) {
     for (uint256 i = 0; i < s_migrationTargets.length; i++) {
       if (s_migrationTargets[i] == target) {
@@ -637,7 +644,9 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
       revert CoordinatorNotRegistered(newCoordinator);
     }
     (uint96 balance, uint96 nativeBalance, , address owner, address[] memory consumers) = getSubscription(subId);
+    // solhint-disable-next-line custom-errors
     require(owner == msg.sender, "Not subscription owner");
+    // solhint-disable-next-line custom-errors
     require(!pendingRequestExists(subId), "Pending request exists");
 
     V1MigrationData memory migrationData = V1MigrationData({
@@ -654,6 +663,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
 
     // Only transfer LINK if the token is active and there is a balance.
     if (address(LINK) != address(0) && balance != 0) {
+      // solhint-disable-next-line custom-errors
       require(LINK.transfer(address(newCoordinator), balance), "insufficient funds");
     }
 
@@ -698,7 +708,7 @@ contract VRFCoordinatorV2PlusUpgradedVersion is
       revert SubscriptionIDCollisionFound();
     }
 
-    for (uint i = 0; i < migrationData.consumers.length; i++) {
+    for (uint256 i = 0; i < migrationData.consumers.length; i++) {
       s_consumers[migrationData.consumers[i]][migrationData.subId] = 1;
     }
